feat: Move global vars to other files
